fix incorrect types
Fixes problems with floats being passed instead of ints,
and timestamp() method being passed instead of the an
actual timestamp.

+++ qwatson/widgets/tableviews.py
@@ -498,7 +498,7 @@ class BasicWatsonTableView(QTableView):
         # ---- Setup column size
 
         self.setColumnWidth(columns['tags'],
-                            1.5 * self.horizontalHeader().defaultSectionSize())
+                            int(1.5 * self.horizontalHeader().defaultSectionSize()))
         self.setColumnWidth(columns['icons'],
                             icons.get_iconsize('small').width() + 12)
         self.horizontalHeader().setSectionResizeMode(QHeaderView.Fixed)
